set modal solver displacements to zero initially.
This would cause problems for the first iteration where the aero is ran before the struct with these
displacements. As a result, the first iteration of aero just gets whatever value is saved in the openmdao
vector for the structural displacements.

Also added comments in the modal solver and tacs components for this.