Updated the modal solver to run on root proc only for parallel runs (#27)
* fixed typo in modal solver

* set modal solver displacements to zero initially.

This would cause problems for the first iteration where the aero is ran before the struct with these
displacements. As a result, the first iteration of aero just gets whatever value is saved in the openmdao
vector for the structural displacements.

Also added comments in the modal solver and tacs components for this.

* Fake parallelizing modal solver. wip

* flattened the mode_shape variable to get the parallelization working. works with meld, doesnt work with RLT yet

* changes to have pseudo parallelized modal solver with RLT

* fixed the comm_self that I forgot to remove.

The code should perform the same way even if the procs that are structural members have the full comm instead of the partial comm.
This is because the python object for RLT handles these cases in the baseTransfer.py file when the structural root is being determined.