Changed the mesh connections behavior.
Now, if no keys exist as funcs or solver, all keys of the dict will be connected to the values under the solver element of that component. If get_mesh _connections provide keys as funcs or solver, the value for this key should be a dictionary itself that defines the connections from the mesh to the functions and solver, respectively.