Include optional functionals group on scenario level. (#29)
* moved all elements to solver_group level in preparation for scenario level elements

* moved adflow funcs to the scenario level.

Also changed the scenario connections output to a dict from a list. This way, we do not need to promote everything we want to connect to the adflow group level. instead, the component can specify the full path and mphys can take care of the stuff it knows.

* moved tacs functionals to scenario level

* updated modal solver with the new mesh connection stuff.

* Changed the mesh connections behavior.

Now, if no keys exist as funcs or solver, all keys of the dict will be connected to the values under the solver element of that component. If get_mesh _connections provide keys as funcs or solver, the value for this key should be a dictionary itself that defines the connections from the mesh to the functions and solver, respectively.

* updated examples with the new hierarchy